move keyboard detection to utils/hooks
