+++ src/core/utils/hooks.ts
@@ -20,3 +20,37 @@ export const useValidBrowser = (keywords?: string[]) => {
 
   return valid;
 };
+
+interface Keyboard {
+  getLayoutMap: () => any;
+}
+
+interface KeyboardLayoutMap {
+  get: (key: string) => any;
+}
+interface Navigator {
+  keyboard: Keyboard;
+}
+
+/**
+ * Check validity of browser to run 3d experiences,
+ * Automatically blacklists Facebook & Instagram in-app
+ * browsers
+ *
+ * @param keywords
+ */
+export const useKeyboardLayout = (keywords?: string[]) => {
+  const [layout, setLayout] = useState("W/A/S/D");
+
+  useEffect(() => {
+    if (navigator.keyboard) {
+      const keyboard = navigator.keyboard;
+      keyboard.getLayoutMap().then((keyboardLayoutMap: KeyboardLayoutMap) => {
+        const upKey = keyboardLayoutMap.get("KeyW");
+        if (upKey === "z") setLayout("Z/Q/S/D");
+      });
+    }
+  }, []);
+
+  return layout;
+};
